Connection to the database
I added the connection to the local database so that I could work on the functionality of the login and sign up page.

The login in page will now allow the users to enter the username or email along with the password for that account. It then checks the provided information against the database and allows the user to continue.

The sign up page will allow the user to enter the required information. It will then enter the provided information into the database.

This does not have a way of authenticating emails and is extremely vulnerable to things like botting and is not intended to be a mainstream application.